+++ vim/pack/plugins/start/fzf/src/tui/light.go
@@ -189,6 +189,10 @@ func (r *LightRenderer) Init() {
 	}
 }
 
+func (r *LightRenderer) Resize(maxHeightFunc func(int) int) {
+	r.maxHeightFunc = maxHeightFunc
+}
+
 func (r *LightRenderer) makeSpace() {
 	r.stderr("\n")
 	r.csi("G")
@@ -540,7 +544,7 @@ func (r *LightRenderer) mouseSequence(sz *int) Event {
 
 	t := atoi(elems[0], -1)
 	x := atoi(elems[1], -1) - 1
-	y := atoi(elems[2], -1) - 1
+	y := atoi(elems[2], -1) - 1 - r.yoffset
 	if t < 0 || x < 0 || y < 0 {
 		return Event{Invalid, 0, nil}
 	}
@@ -676,6 +680,9 @@ func (r *LightRenderer) MaxX() int {
 }
 
 func (r *LightRenderer) MaxY() int {
+	if r.height == 0 {
+		r.updateTerminalSize()
+	}
 	return r.height
 }
 
@@ -855,6 +862,9 @@ func attrCodes(attr Attr) []string {
 	}
 	if (attr & Reverse) > 0 {
 		codes = append(codes, "7")
+	}
+	if (attr & StrikeThrough) > 0 {
+		codes = append(codes, "9")
 	}
 	return codes
 }