+++ vim/pack/plugins/start/vim-lsp-settings/installer/install-volar-server.sh
@@ -1,7 +1,7 @@
-#!/usr/bin/env bash
+#!/bin/sh
 
 set -e
 
-"$(dirname "$0")/npm_install.sh" vue-language-server @volar/vue-language-server@0.37.7
+"$(dirname "$0")/npm_install.sh" vue-language-server @volar/vue-language-server@0.40.13
 mv vue-language-server volar-server
-npm install typescript@4.7
+npm install typescript@4.8